pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "webfaq"
version = "1.0.0"
description = "Creating a Q&A dataset from FAQ-style web page annotations."
authors = [
    "Michael Dinzinger"
]
homepage = "https://www.fim.uni-passau.de"
repository = "https://www.fim.uni-passau.de"
readme = "README.md"
license = "MIT"
packages = [
    {include = "webfaq", from = "src"}
]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
click = "^8.1.7"
# pandas = "^2.2.2"
python-dotenv = "^1.0.1"
huggingface-hub = "^0.28.0"
datasets = "^2.21.0"
# resiliparse = "^0.14.9"
python-iso639 = "^2024.10.22"
torch = "^2.5.1"
transformers = "^4.47.1"
# accelerate = "^1.2.1"
openai = "^2.13.0"
mteb = "2.3.11"
# einops = "^0.8.0"
numpy = "1.26.4"
# setuptools = "^74.1.2"
# wheel = "^0.44.0"
# faiss-cpu = "^1.7.2"
# matplotlib = "^3.10.0"
# seaborn = "^0.13.2"
emoji = "^2.14.0"
# sentence-transformers = "^3.4.0"
# pyserini = "^0.44.0"
isort = "^7.0.0"
black = "^25.12.0"

[tool.poetry.scripts]
webfaq = "webfaq.webfaq:main"
